# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Plateforme E-learning",
        "description": "Documentation API pour la plateforme e-learning",
        "contact": {
            "email": "diopawa425@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Serveur API Local"
        }
    ],
    "paths": {
...
        "/api/lesson/{lesson}/resources": {
            "get": {
                "tags": [
                    "Ressources"
                ],
                "summary": "Lister tous les ressources d'une Leçon",
                "operationId": "7cfe90180008ecb29e2b5ad7a73ebd33",
                "parameters": [
                    {
                        "name": "lesson",
                        "in": "path",
                        "description": "ID de la Leçon",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des ressources",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Ressource"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctumAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Ressources"
                ],
                "summary": "Création de nouvelle ressource pour une Leçon",
                "operationId": "935c16222d1a96398e7773099109c621",
                "parameters": [
                    {
                        "name": "lesson",
                        "in": "path",
                        "description": "ID de la Leçon",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Ressource"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Leçon créé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Ressource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Données invalides"
                    }
                },
                "security": [
                    {
                        "sanctumAuth": []
                    }
                ]
            }
        },
        "/api/lesson/{lesson}/resources/{resource}": {
            "put": {
                "tags": [
                    "Ressources"
                ],
                "summary": "Modifier ou mettre a jour les inlfos d'une Leçon",
                "operationId": "e46ad623443c4ce60a25c9fd31001562",
                "parameters": [
                    {
                        "name": "lesson",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "resource",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Ressource"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Modifier avec succes",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Ressource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Ressource non trougvé"
                    }
                },
                "security": [
                    {
                        "sanctumAuth": []
                    }
                ]
            }
        },
        "/api/resources/{resource}": {
            "delete": {
                "tags": [
                    "Ressources"
                ],
                "summary": "Supprimer une ressource",
                "operationId": "9324268fe414cd2963ba2509763c0d6d",
                "parameters": [
                    {
                        "name": "resource",
                        "in": "path",
                        "description": "ID de la ressource à supprimer",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Leçon supprimé avec succès"
                    },
                    "404": {
                        "description": "Leçon non trouvé"
                    }
                },
                "security": [
                    {
                        "sanctumAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Ressource": {
                "required": [
                    "id",
                    "type"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 2
                    },
                    "title": {
                        "type": "string",
                        "example": "Livre Laravel"
                    },
                    "type": {
                        "type": "string",
                        "example": "pdf"
                    },
                    "description": {
                        "type": "string",
                        "nullable": true
                    },
                    "path": {
                        "type": "string",
                        "example": "lessons/files/example.pdf"
                    },
                    "external_url": {
                        "type": "string",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctumAuth": {
                "type": "http",
                "description": "Entrez le token Sanctum reçu après connexion.",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}